overo: handle board revision detection for fab revisions <= R2410
Revisions < R2410 connect gpio112 to the 4030 irq_1 line. Since gpio112
is now used for board revision, this could lead to an erroneous
board revision results depending on the state of the irq.

This patch forces a 4030 RTC interrupt, which pulls irq_1 low, giving
a low level on gpio112 on these boards.  This gives a proper revision
0 result for these boards.